bug fixed

+++ illust_repo.py
@@ -121,7 +121,7 @@ class IllustratorRepos(object):
         if nbr <= 20:
             needPagecnt = 1                                         # nbr <= 20, request once
         else:
-            needPagecnt = (nbr / 20) + 1                            # calcus need request count
+            needPagecnt = int(nbr / 20) + 1                         # calcus need request count
 
         # gather all data(thumbnail images and names)
         allThumbnailimage = []
